# Performance optimizations for FormatParser